Refaktor: InMemory- és JpaPizzaService közös elemek kiemelése
