+++ PizzaApp/src/main/java/com/github/holiver98/service/inmemory/InMemoryPizzaService.java
@@ -7,11 +7,12 @@ import com.github.holiver98.model.Ingredient;
 import com.github.holiver98.model.Pizza;
 import com.github.holiver98.model.Rating;
 import com.github.holiver98.service.IPizzaService;
+import com.github.holiver98.service.PizzaServiceBase;
 
 import java.math.BigDecimal;
 import java.util.List;
 
-public class InMemoryPizzaService implements IPizzaService {
+public class InMemoryPizzaService extends PizzaServiceBase {
 
     private IInMemoryIngredientDao ingredientDao;
     private IInMemoryPizzaDao pizzaDao;
@@ -24,60 +25,8 @@ public class InMemoryPizzaService implements IPizzaService {
     }
 
     @Override
-    public float calculatePrice(Pizza pizza) {
-        if(pizza == null){
-            return -1f;
-        }
-
-        if(pizza.getIngredients() == null){
-            System.out.println("Ingredients list not initialized!");
-            return -1f;
-        }
-
-        if(pizza.getIngredients().size() == 0){
-            System.out.println("There are no ingredients assigned to this pizza!");
-            return -1f;
-        }
-
-        if(!arePricesValid(pizza)){
-            System.out.println("Invalid ingredient prices!");
-            return -1f;
-        }
-
-        BigDecimal totalPrice = new BigDecimal(0);
-        for (Ingredient i: pizza.getIngredients()) {
-            BigDecimal ingredientPrice = new BigDecimal(Float.toString(i.getPrice()));
-            totalPrice = totalPrice.add(ingredientPrice);
-        }
-
-        return totalPrice.floatValue();
-    }
-
-    @Override
-    public float recalculateRatingAverage(long pizzaId) {
-        Pizza pizzaToUpdate = pizzaDao.getPizzaById(pizzaId);
-        if(pizzaToUpdate == null){
-            System.out.println("Pizza with id("+ pizzaId +") was not found in the database!");
-            return -1;
-        }
-
-        List<Rating> ratings = ratingDao.getRatingsOfPizza(pizzaId);
-        int ratingSumm = 0;
-        for (Rating rating : ratings) {
-            ratingSumm += rating.getRating();
-        }
-
-        float newRatingAverage;
-        if(ratings.size() == 0){
-            return 0;
-        }else{
-            newRatingAverage = (float)ratingSumm / (float)ratings.size();
-        }
-
-        pizzaToUpdate.setRatingAverage(newRatingAverage);
-        pizzaDao.updatePizza(pizzaToUpdate);
-
-        return newRatingAverage;
+    protected List<Rating> getRatingsOfPizza(long pizzaId) {
+        return ratingDao.getRatingsOfPizza(pizzaId);
     }
 
     @Override
@@ -124,27 +73,5 @@ public class InMemoryPizzaService implements IPizzaService {
         }
 
         pizzaDao.deletePizza(pizzaId);
-    }
-
-    private boolean arePricesValid(Pizza pizza) {
-        for (Ingredient i: pizza.getIngredients()) {
-            if(i.getPrice() < 0f){
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    private boolean isValidPizza(Pizza pizza){
-        if(pizza == null){
-            return false;
-        }else if(pizza.getId() < 0){
-            return false;
-        }else if(!arePricesValid(pizza)){
-            return false;
-        }
-
-        return true;
     }
 }
